Change GP to use decimal values

+++ src/main/java/org/skytemple/altaria/features/reputation/ChangeGpInterfaceCommand.java
@@ -37,7 +37,7 @@ public class ChangeGpInterfaceCommand extends ChangeGpCommand {
 	 * @param privateResultSender Used to send the message with the result of the command in JSON format
 	 * @param errorSender Used to send error messages
 	 */
-	public ChangeGpInterfaceCommand(ReputationDB rdb, User user, int amount, MessageSender publicResultSender,
+	public ChangeGpInterfaceCommand(ReputationDB rdb, User user, double amount, MessageSender publicResultSender,
 									MessageSender privateResultSender, MessageSender errorSender) {
 		super(rdb, user, amount, null, errorSender);
 		this.publicResultSender = publicResultSender;
