+++ src/main/java/org/skytemple/altaria/features/reputation/GiveGpCommand.java
@@ -30,7 +30,7 @@ public class GiveGpCommand extends ChangeGpCommand {
 	 * @param resultSender Used to send result messages to the user
 	 * @param errorSender Used to send error messages to the user
 	 */
-	public GiveGpCommand(ReputationDB rdb, User user, int amount, MessageSender resultSender, MessageSender errorSender) {
+	public GiveGpCommand(ReputationDB rdb, User user, double amount, MessageSender resultSender, MessageSender errorSender) {
 		super(rdb, user, amount, resultSender, errorSender);
 	}
